% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.data.R
\docType{data}
\name{rf.data}
\alias{rf.data}
\title{Example Random Forest Data}
\format{
A data frame with 2 columns:
\describe{
  \item{actual}{Actual values}
  \item{predicted}{Predicted probabilities}
}
}
\source{
Full test data set using random forest classifier, see Section 6 in Reference.
}
\usage{
rf.data
}
\description{
This dataset contains example data from a Random Forest model.
}
\references{
Holzmann, H., Klar, B. (2024). Robust performance metrics for imbalanced classification problems.
arXiv:2404.07661. \href{https://arxiv.org/abs/2404.07661}{LINK}
}
\keyword{datasets}
